Added form for job creation

+++ app/database/Job.scala
@@ -12,7 +12,7 @@ import slick.driver.MySQLDriver.api._
 
 import java.sql.Date
 
-case class Job(name: String, description:String, startDate: Date, endDate:Date,jobType:Int, region:Int, hourlyPay:Float, workingTime:Int, email:String, id: Option[Int] = None)
+case class Job(name: String, description:String, startDate: Date, endDate:Date,jobType:Int, region:Int, hourlyPay:Double, workingTime:Int, email:String, id: Option[Int] = None)
 
 class Jobs(tag: Tag) extends Table[Job](tag, "jobs") {
   def id = column[Int]("id", O.PrimaryKey, O.AutoInc)
@@ -22,7 +22,7 @@ class Jobs(tag: Tag) extends Table[Job](tag, "jobs") {
   def endDate = column[Date]("endDate")
   def jobType = column[Int]("jobType")
   def region = column[Int]("region")
-  def hourlyPay = column[Float]("hourlyPay")
+  def hourlyPay = column[Double]("hourlyPay")
   def workingTime = column[Int]("workingTime")
   def email = column[String]("email")
   def typeId = foreignKey("typeId", jobType, Types.typeTable)(_.typeId)
@@ -36,7 +36,7 @@ object Jobs {
   val jobTable = TableQuery[Jobs]
 
 
-  def insert(name: String,description:String, startDate: Date, endDate: Date, jobType:Int, region:Int, hourlyPay:Float, workingTime:Int, email:String) =
+  def insert(name: String,description:String, startDate: Date, endDate: Date, jobType:Int, region:Int, hourlyPay:Double, workingTime:Int, email:String) =
     Await.result(db.run(DBIO.seq(jobTable += Job(name, description, startDate, endDate,jobType, region, hourlyPay, workingTime, email ))), Duration.Inf)
 
   def all(): List[Job] = (for (j <- Await.result(db.run(jobTable.result), Duration.Inf)) yield j).toList
